chore: set release to false in projen (#3)
Currently there is an [issue](https://github.com/jenkinsci/kubernetes-operator/pull/729) in jenkins-operator which is causing the regex to fail when we try to update a base plugin in the [jenkins instance manifest](https://jenkinsci.github.io/kubernetes-operator/docs/getting-started/latest/customizing-jenkins/). So setting `release: false` till Jenkins team updates the jenkins-operator image and has an update in the [documentation](https://jenkinsci.github.io/kubernetes-operator/docs/getting-started/latest/installing-the-operator/) as well for the users.

Signed-off-by: Vinayak Kukreja <vinakuk@amazon.com>